refresh token implemented

+++ app/main/resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from ..models.item import ItemModel
 
 
@@ -23,6 +23,7 @@ class Item(Resource):
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -38,7 +39,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['isAdmin']:
+            return {'message': 'Admin privilege required'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -61,7 +67,11 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        # return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
-        # return  {'items': [x.json() for x in ItemModel.query.all()]}
-        return {'items': [x.json() for x in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items = [x.json() for x in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}
+        return {'items': [x['name'] for x in items],
+                'message': 'More Data would be available if u were Admin'}, 200
